chore(ci): fix update separate
Regression caused by the change of `release-type` in
release-please-config-beta.json to be `node` as well, both the beta and
stable release-please configs generate the same branch name when
creating pull requests, and end up clobbering each other. Manually set
the package's `component` so that the branch name that is generated is
differentiated between the two release-please workflows, and
release-please will maintain two separate pull requests.